+++ src/app/layout.tsx
@@ -1,3 +1,4 @@
+import ConditionalLayout from "@/components/layout/conditional-layout";
 import { Footer } from "@/components/layout/footer";
 import { Navbar } from "@/components/layout/navbar";
 import { AuthProvider } from "@/components/providers/auth-provider";
@@ -29,11 +30,9 @@ export default function RootLayout({
           disableTransitionOnChange
         >
           <AuthProvider>
-            <div className="flex min-h-screen flex-col">
-              <Navbar />
-              <main className="flex-1 pt-16">{children}</main>
-              <Footer />
-            </div>
+            <ConditionalLayout navbar={<Navbar />} footer={<Footer />}>
+              {children}
+            </ConditionalLayout>
           </AuthProvider>
         </ThemeProvider>
       </body>
